refactoring-rightSide
